[tests] Remove cancelled output excluded test
This test was broken due to changes from #2644, where we now enforce
unique commitments in the outputs table.

Previously, it tested that cancelled tx outputs with the same commitment
would not be spent instead of the valid output (issue #2457).

This is now solved in 2 ways:
- coinbase outputs are reused, instead of being regenerated with the
  same commitment (#2478)
- unique commitments are enforced in the outputs table (#2644)

Given that our invariants have changed, this test is no longer
required.