[base-node] Return block hash from submit_block call
The block hash is now returned from submit block.

The block hash needs to be calculated from the pow data of the submitted block.
When merge mining, this is not known and needs to be seperately
calculated. The merge mining proxy, for instance, can now return the aux
chain's final hash for monerod submit_block.