Update Cucumber stress test to be a Scenario Outline
This PR updates the cucumber stress test into a Scenario Outline that test sending 10, 100, 1000 and 10000 transactions. Some small fixes were required to get the stress test to work including plumbing in the `flood_ban_max_msg_count` config setting so that the wallet and base nodes don’t ban each other for message flooding during a cucumber test.

Another problem encountered was that when you send 10_000 transactions when the broadcast monitoring timeout is low, which is desired in other cucumber tests, the CPU load when we read the last transactions is very high so this PR include the ability to create a “stress-test” wallet that has a long broadcast monitoring timeout than a standard wallet.

Currently the stress test can send all 10_000 transactions but at that stage the base nodes ban each other when they are unable to sync headers due to RPC concurrency limits so the mined blocks no longer propagate.