[Documentation] Minor Fixes
Fix solo merged mining example config that was mistakenly using a subaddress instead of normal address for mainnet configuration, ModeroD doesn't currently support requesting block templates using subaddresses. Self-select configuration can remain the same as the pool wallet address is used to request the template, the miner address is purely for the pool to be able to pay out the miner.

Updated docs to reflect the same.

Review comments