[wallet] Differentiate "funds pending" case in UTXO selection
Adds a new error case to the output manager service. When selecting
UTXOs and not enough spendable funds are available, check if the pending
incoming balance would cover the shortfall. If so, return the new
`FundsPending` error instead of `NotEnoughFunds`. This way, the caller
can differentiate between the two cases.