Implement wallet recovery process in LibWallet FFI
This PR adds the functionality to the LibWallet FFI to recovery a wallet from a set of seed words. 

There was an implementation of the recovery process in the Tari Console Wallet and this was moved into LibWallet and generalised to be used for both clients. The process is now started and a stream of events will be used to inform the client of the progress. Furthermore the progress of the process is now persisted in the LibWallet db backend to allow for the resuming of the process from where it left off. The Console wallet was updated to check if a restoration is in progress and to continue it.

- Added the ability to create a TariSeedWordsObject and add words to it. The u8 enum return from `seed_words_push_word(…)` tells the client if the word was valid and when the seed phrase of 24 words is complete.
- Add a call to check if there was an In Progress restoration. The wallet client can ignore this if they want to, the wallet functions as normal but obviously we won’t be sure that all outputs have been recovered or not.
- Add a call to start the recovery process. The client provides a callback whose arguments are what block we have recovered to and the current total chain length. Based on these two values the client will be informed of the status of the asynchronous recovery process.