Updated `/// Sum of script offsets for all kernels in this block.`
Implement TariScript base node validation and default scripts in transactions

This PR includes the work of @SWvheerden, @hansieodendaal and myself to implement the foundations of TariScript support in the Tari base node and wallets.

The goal of this PR is to add all the data required for TariScripts while building transactions and also to validate the TariScripts by base nodes. 

On the wallet side of things this PR updates the wallet Sql database to store all the elements that the wallet needs to remember to spend an output with a script and updates the Transaction Negotiation Protocols to work with TariScripts. The updates are able to support general TariScripts but in the Wallet the current assumption is that all outputs will have the “Default” Nop script for now.

This PR also adds the fields to the headers required for script validation and the base node validation has been updated to also check the script elements for individual transactions and the accumulated offset for  constructed blocks and the blockchain.

One last feature that is included in this PR as it was easy to add in during the Transaction Protocol refactor is that the Coin Split function in the Wallet now produces outputs that are rewindable and thus can be recovered by a wallet recovery.

fix unit tests

Some code review comments