[Wallet] Fix wallet service API loop blocking
It was observed that during a stress test which fires a series of send_transaction calls in the wallet that there was a significant delay between some calls up to 30 seconds long during which no other Transaction Service events were firing in the services main select! loop. This turned out to be because the `send_transaction `operation involved a call to the Base Node Service (BNS) to check the latest tip height and this service was blocking its select! Loop when it made RPC calls to the base node. If these RPC calls took long the BNS service would not respond to API calls until The RPC request was complete. This blocked the select loops of the Output Manager Service and Transaction Service that relied on this call being fast. This also impacted the `get_balance` operation from the Output Manager Service which relies on a check of the last known tip height.

To fix this the BNS service was refactored so that the GetChainMetadata monitoring occurs in a separate asynchronous task that updates the services state when it gets an answer via RPC. The main select! loop was thus freed up to only respond to API requests that provide either the last updated chain metadata OR if no contact has been made with a base node during this run it will get the last persisted data from the database.

Another improvement that was made for the make-it-rain operation in the console wallet was to prevent the console wallet from performing the UTXO and Transaction validation operations when being run in `script` or `command` mode. The validation operations are long running tasks and should only be run if the console wallet is being run in a long running TUI mode.