[Wallet] Update wallet gRPC server to send one-sided payments
This PR adds an enum to the PaymentRecipient Protocol struct that is used with the Transfer method to allow the user to specify if a payment is made using the traditional Mimblewimble method or as a One-sided payment.

This was tested manually as the Cucumber tests seem broken in the TariScript branch currently.