Emission math audit
Check for overflows with integer math in emission schedule and adds
tests.

Shift right on u64 will overflow if shifting 64 or more bits. The shift
is ultimately controlled by hard-coded consensus constants. This PR adds
a panic if the emission decay is set too high (i.e. >=64). In release, an
shift right overflow is undefined, meaning that in the unlikely event that
a hard-coded value is set incorrectly, it could cause an inflation bug.