Implement including full script in transaction outputs
This PR replaces the script hash in a transaction output with a full script. This has been done to make on-chain script detection and analysis possible. This makes features like detecting your own scripts based on included pubkeys during a seed phrase recovery possible.

This PR updates the TransactionOutput struct to hold a script and not script hash and updates the code base to reflect this change.

RFC-201 is updated to reflect the change to the calculation of Beta