Cancel faux transaction when imported UTXO is invalidated
When a UTXO is imported into the wallet a faux transaction is added to the transaction service to represent the event that impacted the wallets balance. If the imported output turns out to be invalid, i.e. it was an already spent output or it was an orphan output then the output validation will invalidate it. However, the faux transaction would not be invalidated and would still appear in wallet transaction lists. This was found to be confusing. So this PR links the imported output to the faux transaction and if the output is invalidated the transaction is invalidated.

Cucumber tests are added to test this functionality.