Remove MemoryDBs from the wallet
In the past all the wallet services has a Sqlite and Memory based db backend to use in testing. However maintaining the memory DB’s became a chore.  This PR removes the Output Manager and Contacts Service memory databases and updates the tests that use those databases to use a temporary Sqlite database instead.