Merge #3018: [tests] Fix block_event_and_reorg_event_handling test
This PR fix 3 issues that were failing:

The coinbase_lock_height is 2, so we had to add one empty block to be able to
process the UTXO.  The TXs were discarded because their input was invalid.
Solution was to move it after the processing of the block1.  Removal of
assert_eq, there is proper one async_assert_eventually.