chore: Upgrade yamux v0.9.0
Upgrade `yamux` library used for TCP stream muxing in comms to 0.9.0.
This upgrade contains a number of [changes and
fixes](https://github.com/libp2p/rust-yamux/blob/develop/CHANGELOG.md) that should be
incorporated.
Increased the recieve window size to be larger than the maximum RPC
message size. Anecdotally, since RPC messages are allowed to be 4Mib
which happens to be the size of the yamux window, large messages will
exhaust the entire window which is not optimial. The deadlock case doesn't
make sense because the RPC protocol never writes at the same time (strictly
alternating between read/write).