fix: improve reliability of get block template protocol in mm proxy
The main change here is to check the tip height after retrieving the
block template but before MMR roots are calculated. When mining quickly
in cucumber tests, this race condition occurs. Previously, this would be
ok, the miner would simply mine 1 behind the current tip height. But
because we can no longer calculate MMRs prior to the tip, this MMR call
(`get_new_block`) now returns an error.

Emits handling errors to stderr to make it quicker/easier to diagnose issues
without going into the logs.

This refactor was performed because the protocol was gaining too much
complexity to exist in a single function.