feat: add `ping()` to all comms RPC clients
Adds a `ping()` function to all comms RPC clients. `ping()` sends an RPC
request with the `ACK` flag set. The server will immediately reply with
an `ACK` response. This accurately measures RPC latency without a
potentially slow backend. `ping()` is now used in the wallet monitor.

Previously, `get_last_request_latency` was used which will increase
whenever the blockchain db is busy, for e.g:
- the base node is syncing
- another node(s) syncing from the base node
- one or many wallets scanning UTXOs from the base node
- one or many wallets running a recovery from the base node
- large reorg

A client wallet would, of course, have no idea that this is occurring
and simply display poor latency. This could be perceived to be a poor
RPC performance, when in fact, there are a number or non-network/RPC
related reasons why a multi-second ping is displayed.

`get_last_request_latency` is a better measure when determining
base node performance (depending on the RPC method impl). However,
it is misleading to use this as a user-facing value presented as network latency.