feat: add logging of Monero PoW data to debug merge mining
An intermittent error is being observed where Xmrig submits what it considers a valid PoW solution but the Base Node validation rejects the block with a low (1, 2, 3) difficulty.

Adding two sets of logs to provide more detail on the data used to perform the RandomX hash when verifying a submitting merge mined block. Added a log to the  Merge Mining Proxy to show the full Monero PoW data received from Xmrig and added a log in the Base Node to show the RandomX hash that it produces.

These logs will help us confirm where the issue is in this disconnect between Xmrig’s calculation and our validation.