fix: resolved design flaw in wallet_ffi library
Exposed TariTransactionKernel instead of TariExcess, TariExcessSignature and TariExcessPublicNonce. This was done as they are all referring to fields from the same object and also to reduce the amount of wrapper classes needs on the other side of the FFI boundary.

Added accessor methods to retrieve data from the opaque pointer for TariTransactionKernel and return the data as a type (in this instance a cstring) which can safely cross the FFI boundary.

Updated library header.

Updated rust tests.

Updated cucumber tests.

Updated comments.

Incremented library version.

Fix typo

Fix duplicate line