feat: update+notifications for console wallet (#3284)
Add general notifications support for console wallet.
Add update notification to the wallet.

Description
---
The wallet now supports notifications. In case of notification the _notification_ tab will change color and show number of notifications so the users knows there is something. Once you sees them the counters goes to zero and the color to white.

Motivation and Context
---
It's the preparation for the auto update feature. Currently it's just notification for the update availability.

How Has This Been Tested?
---
The general notifications was tested by adding notification from the base node change.
The auto update was not tested yet.