chore: change CI timeout to 20 minutes (#3289)
Description
---
Change the CI timeout on cucumber tests to 20 minutes.
In case a CI test will break still inside the tests (and not in the after step). We should investigate the test and not ignore the timeout kill.

Motivation and Context
---
Most of the long steps in the cucumber test have 20 minutes timeout. So if they by any chance don't output anything they will not be killed after 10 minutes (current scenario).
We still have some longer tests. The longest timeout is 12 hours.